+++ packages/web3-hooks/base/src/index.ts
@@ -72,8 +72,6 @@ export * from './useRiskWarningApproved.js'
 export * from './useSchemaType.js'
 export * from './useSocialAccounts.js'
 export * from './useSocialAccountsAll.js'
-export * from './useSocialAddressList.js'
-export * from './useSocialAddressListAll.js'
 export * from './useTransactions.js'
 export * from './useTrustedFungibleTokens.js'
 export * from './useTrustedNonFungibleTokens.js'
