+++ packages/web3-hooks/base/src/useSocialAccountsAll.ts
@@ -1,7 +1,7 @@
 import { EMPTY_LIST, NetworkPluginID } from '@masknet/shared-base'
 import type { SocialAddress, SocialAddressType, SocialIdentity } from '@masknet/web3-shared-base'
 import { useSocialAddressListAll } from './useSocialAddressListAll.js'
-import { useSocialAccounts } from './useSocialAccounts.js'
+import { useSocialAccountsFrom } from './useSocialAccountsFrom.js'
 
 /**
  * Get all social addresses under of all networks.
@@ -12,7 +12,7 @@ export function useSocialAccountsAll(
     sorter?: (a: SocialAddress<NetworkPluginID>, z: SocialAddress<NetworkPluginID>) => number,
 ) {
     const { value: socialAddressList = EMPTY_LIST, ...rest } = useSocialAddressListAll(identity, includes, sorter)
-    const socialAccounts = useSocialAccounts(socialAddressList)
+    const socialAccounts = useSocialAccountsFrom(socialAddressList)
 
     return {
         ...rest,
